+++ test/clj_kondo/redundant_nested_call_test.clj
@@ -4,19 +4,19 @@
 
 (deftest redundant-nesting-test
   (assert-submaps2
-   '({:file "<stdin>", :row 1, :col 10, :level :warning, :message "Redundant nested call: min"})
+   '({:file "<stdin>", :row 1, :col 10, :level :info, :message "Redundant nested call: min"})
    (lint! "(min 5 2 (min 3 7))"))
   (assert-submaps2
-   '({:file "<stdin>", :row 1, :col 8, :level :warning, :message "Redundant nested call: *"})
+   '({:file "<stdin>", :row 1, :col 8, :level :info, :message "Redundant nested call: *"})
    (lint! "(* 3 4 (* 5 6))"))
   (assert-submaps2
-   '({:file "<stdin>", :row 1, :col 18, :level :warning, :message "Redundant nested call: str"})
+   '({:file "<stdin>", :row 1, :col 18, :level :info, :message "Redundant nested call: str"})
    (lint! "(str \"foo\" \"bar\" (str \"baz\" \"qux\"))"))
   (assert-submaps2
-   '({:file "<stdin>", :row 1, :col 21, :level :warning, :message "Redundant nested call: concat"})
+   '({:file "<stdin>", :row 1, :col 21, :level :info, :message "Redundant nested call: concat"})
    (lint! "(concat [1 2] [3 4] (concat [5 6] [7 8]))"))
   (assert-submaps2
-   '({:file "<stdin>", :row 1, :col 21, :level :warning, :message "Redundant nested call: concat"})
+   '({:file "<stdin>", :row 1, :col 21, :level :info, :message "Redundant nested call: concat"})
    (lint! "(concat [1 2] [3 4] (concat [5 6] [7 8]))"))
   (is
    (empty?