+++ CHANGELOG.md
@@ -19,14 +19,15 @@ For a list of breaking changes, check [here](#breaking-changes).
 
 ## Unreleased
 
+- [#2212](https://github.com/clj-kondo/clj-kondo/issues/2212): NEW linter: `:redundant-nested-call` ([@tomdl89](https://github.com/tomdl89)), set to level `:info` by default
+- Bump `:redundant-ignore`, `:redundant-str-call` linters to level `:info`
 - [#1784](https://github.com/clj-kondo/clj-kondo/issues/1784): detect `:redundant-do` in `catch`
 - [#2410](https://github.com/clj-kondo/clj-kondo/issues/2410): add `--report-level` flag
 - [#2416](https://github.com/clj-kondo/clj-kondo/issues/2416): detect empty `require` and `:require` forms ([@NoahTheDuke](https://github.com/NoahTheDuke))
 - [#1786](https://github.com/clj-kondo/clj-kondo/issues/1786): Support `gen-interface` (by suppressing unresolved symbols)
 - [#2407](https://github.com/clj-kondo/clj-kondo/issues/2407): support ignore hint on called symbol
-- [#2420](https://github.com/clj-kondo/clj-kondo/issues/2420): Detect uneven number of clauses in `cond->` and `cond->>`
+- [#2420](https://github.com/clj-kondo/clj-kondo/issues/2420): Detect uneven number of clauses in `cond->` and `cond->>` ([@tomdl89](https://github.com/tomdl89))
 - [#2415](https://github.com/clj-kondo/clj-kondo/issues/2415): false positive type checking issue with `str/replace` and `^String` annotation
-- [#2212](https://github.com/clj-kondo/clj-kondo/issues/2212): new linter: `:redundant-nested-call` ([@tomdl89](https://github.com/tomdl89))
 
 ## 2024.09.27